+++ spec/logstash/outputs/scalyr_spec.rb
@@ -61,6 +61,7 @@ describe LogStash::Outputs::Scalyr do
     context "test get_stats and send_status" do
       plugin = LogStash::Outputs::Scalyr.new({
                                                      'api_write_token' => '1234',
+                                                     'perform_connectivity_check' => false,
                                                      'serverhost_field' => 'source_host',
                                                      'log_constants' => ['tags'],
                                                      'flatten_nested_values' => true,
@@ -77,6 +78,7 @@ describe LogStash::Outputs::Scalyr do
       it "it doesnt include flatten metrics if flattening is disabled" do
         plugin1 = LogStash::Outputs::Scalyr.new({
                                                      'api_write_token' => '1234',
+                                                     'perform_connectivity_check' => false,
                                                      'serverhost_field' => 'source_host',
                                                      'log_constants' => ['tags'],
                                                      'flatten_nested_values' => false,
@@ -151,6 +153,7 @@ describe LogStash::Outputs::Scalyr do
       it "send_stats is not called when events list is empty and report_status_for_empty_batches is false" do
         plugin2 = LogStash::Outputs::Scalyr.new({
                                                      'api_write_token' => '1234',
+                                                     'perform_connectivity_check' => false,
                                                      'serverhost_field' => 'source_host',
                                                      'log_constants' => ['tags'],
                                                      'flatten_nested_values' => false,
@@ -206,6 +209,7 @@ describe LogStash::Outputs::Scalyr do
       it "creates logfile from serverHost" do
         plugin = LogStash::Outputs::Scalyr.new({
                                                    'api_write_token' => '1234',
+                                                   'perform_connectivity_check' => false,
                                                    'serverhost_field' => 'source_host',
                                                    'log_constants' => ['tags'],
                                                })
@@ -223,7 +227,7 @@ describe LogStash::Outputs::Scalyr do
 
     context "when serverhost_field is missing" do
       it "does not contain log file" do
-        plugin = LogStash::Outputs::Scalyr.new({'api_write_token' => '1234'})
+        plugin = LogStash::Outputs::Scalyr.new({'api_write_token' => '1234', 'perform_connectivity_check' => false})
         allow(plugin).to receive(:send_status).and_return(nil)
         plugin.register
         result = plugin.build_multi_event_request_array(sample_events)
@@ -237,6 +241,7 @@ describe LogStash::Outputs::Scalyr do
       it "creates logfile from serverHost" do
         plugin = LogStash::Outputs::Scalyr.new({
                                                    'api_write_token' => '1234',
+                                                   'perform_connectivity_check' => false,
                                                    'serverhost_field' => 'source_host',
                                                })
         allow(plugin).to receive(:send_status).and_return(nil)
@@ -254,6 +259,7 @@ describe LogStash::Outputs::Scalyr do
       it "does not contain log file" do
         plugin = LogStash::Outputs::Scalyr.new({
                                                    'api_write_token' => '1234',
+                                                   'perform_connectivity_check' => false,
                                                    'serverhost_field' => 'source_host',
                                                    'logfile_field' => 'source_file',
                                                })
@@ -271,6 +277,7 @@ describe LogStash::Outputs::Scalyr do
     context "when configured to flatten values with custom delimiter" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
           'flatten_tags' => true,
           'flat_tag_value' => 'true',
           'flat_tag_prefix' => 'tag_prefix_',
@@ -303,6 +310,7 @@ describe LogStash::Outputs::Scalyr do
     context "when configured to flatten values with custom delimiter and deep delimiter fix" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
           'flatten_tags' => true,
           'flat_tag_value' => 'true',
           'flat_tag_prefix' => 'tag_prefix_',
@@ -336,6 +344,7 @@ describe LogStash::Outputs::Scalyr do
     context "when configured to flatten values with custom delimiter, no array flattening" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
           'flatten_tags' => true,
           'flat_tag_value' => 'true',
           'flat_tag_prefix' => 'tag_prefix_',
@@ -367,6 +376,7 @@ describe LogStash::Outputs::Scalyr do
     context "when configured to flatten values and tags" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
           'flatten_tags' => true,
           'flat_tag_value' => 'true',
           'flat_tag_prefix' => 'tag_prefix_',
@@ -399,6 +409,7 @@ describe LogStash::Outputs::Scalyr do
       it "estimate_each_event_size is true explicit (default) batch split into 3 scalyr requests" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'flatten_tags' => true,
             'flat_tag_value' => 'true',
             'flat_tag_prefix' => 'tag_prefix_',
@@ -439,6 +450,7 @@ describe LogStash::Outputs::Scalyr do
       it "estimate_each_event_size is true implicit (default) batch split into 3 scalyr requests" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'flatten_tags' => true,
             'flat_tag_value' => 'true',
             'flat_tag_prefix' => 'tag_prefix_',
@@ -478,6 +490,7 @@ describe LogStash::Outputs::Scalyr do
       it "estimate_each_event_size is false batch not split into multiple scalyr requests" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'flatten_tags' => true,
             'flat_tag_value' => 'true',
             'flat_tag_prefix' => 'tag_prefix_',
@@ -513,6 +526,7 @@ describe LogStash::Outputs::Scalyr do
     context "when not configured to flatten values and tags" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
       }
       plugin = LogStash::Outputs::Scalyr.new(config)
       it "does not flatten" do
@@ -535,6 +549,7 @@ describe LogStash::Outputs::Scalyr do
     context "when configured to flatten with max keys configured to 3" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
           'flatten_nested_values' => true,  # this converts into string 'true'
           'flattening_max_key_count' => 3,
       }
@@ -567,6 +582,7 @@ describe LogStash::Outputs::Scalyr do
       it "no serverHost defined in server_attributes, no serverHost defined on event level - should use node hostname as the default session level value" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
 
@@ -585,6 +601,7 @@ describe LogStash::Outputs::Scalyr do
       it "serverHost defined in server_attributes, nothing defined on event level - server_attributes value should be used" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'serverHost' => 'fooHost'}
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -604,6 +621,7 @@ describe LogStash::Outputs::Scalyr do
       it "serverHost defined in server_attributes (explicitly defined), event level serverHost defined - event level value should be used" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'serverHost' => 'fooHost', 'attr1' => 'val1'}
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -660,6 +678,7 @@ describe LogStash::Outputs::Scalyr do
       it "serverHost defined in server_attributes (defined via node hostname), event level serverHost defined - event level value should be used" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'attr1' => 'val1'}
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -719,6 +738,7 @@ describe LogStash::Outputs::Scalyr do
       it "serverHost defined in server_attributes (explicitly defined), event level serverHost defined - event level value should be used and server level one for events without server host" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'serverHost' => 'top-level-session-host', 'attr1' => 'val1'}
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -771,6 +791,7 @@ describe LogStash::Outputs::Scalyr do
       it "no serverHost defined, event level serverHost defined via non-default serverhost_field - event level value should be used" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'attr1' => 'val1'},
             'use_hostname_for_serverhost' => false,
             'serverhost_field' => 'custom_server_host',
@@ -827,6 +848,7 @@ describe LogStash::Outputs::Scalyr do
       it "no serverHost defined, event level serverHost defined - event level value should be used" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'server_attributes' => {'attr1' => 'val1'},
             'use_hostname_for_serverhost' => false
         }
@@ -882,6 +904,7 @@ describe LogStash::Outputs::Scalyr do
     context "when receiving an event with Bignums" do
       config = {
           'api_write_token' => '1234',
+          'perform_connectivity_check' => false,
       }
       plugin = LogStash::Outputs::Scalyr.new(config)
       it "doesn't throw an error" do
@@ -902,6 +925,7 @@ describe LogStash::Outputs::Scalyr do
       it "host attribute removed by default" do
        config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
 
@@ -927,6 +951,7 @@ describe LogStash::Outputs::Scalyr do
       it "host attribute not removed if config option set" do
        config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'remove_host_attribute_from_events' => false,
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -955,6 +980,7 @@ describe LogStash::Outputs::Scalyr do
       it "stdlib (implicit)" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
 
@@ -971,6 +997,7 @@ describe LogStash::Outputs::Scalyr do
       it "stdlib (explicit)" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'json_library' => 'stdlib'
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -988,6 +1015,7 @@ describe LogStash::Outputs::Scalyr do
       it "jrjackson" do
         config = {
             'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
             'json_library' => 'jrjackson'
         }
         plugin = LogStash::Outputs::Scalyr.new(config)
@@ -1000,6 +1028,66 @@ describe LogStash::Outputs::Scalyr do
         body = JSON.parse(result[0][:body])
         expect(result[0][:body]).to include(sprintf('{"serverHost":"%s","monitor":"pluginLogstash"}', NODE_HOSTNAME))
         expect(body['events'].size).to eq(1)
+      end
+    end
+
+    context "scalyr_server config option handling and connectivity check" do
+      it "doesn't throw an error on valid url" do
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'https://agent.scalyr.com'
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        plugin.register
+
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'https://eu.scalyr.com'
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        plugin.register
+      end
+
+      it "throws on invalid URL" do
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'agent.scalyr.com'
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        expect { plugin.register }.to raise_error(LogStash::ConfigurationError, /is not a valid URL/)
+
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'eu.scalyr.com'
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        expect { plugin.register }.to raise_error(LogStash::ConfigurationError, /is not a valid URL/)
+      end
+
+      it "throws on invalid hostname" do
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'https://agent.invalid.foo.scalyr.com',
+            'perform_connectivity_check' => true
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        expect { plugin.register }.to raise_error(LogStash::ConfigurationError, /Received error when trying to communicate/)
+      end
+
+      it "throws on invalid api key" do
+        config = {
+            'api_write_token' => '1234',
+            'perform_connectivity_check' => false,
+            'scalyr_server' => 'https://agent.scalyr.com',
+            'perform_connectivity_check' => true
+        }
+        plugin = LogStash::Outputs::Scalyr.new(config)
+        expect { plugin.register }.to raise_error(LogStash::ConfigurationError, /Received 401 from Scalyr API during/)
       end
     end
   end